Improve integration of new platforms. Close #72
SPHDE has to know the pointer word size and platform Endian.
This was simple for 1st two platforms (PowerPC and X86) but got to
be a bit of a cludge as more platforms are added. So want sasconf.h
to establish __WORDSIZE__64/32 and let the compiler provide the
target Endian:
 - (__BYTE_ORDER__ == __ORDER_LITTLE_ENDIAN__)
 - (__BYTE_ORDER__ == __ORDER_BIG_ENDIAN__)

 Everywhere else use these defines to size and order bit fields.

 	* src/sasanchr.h (regionFlags[__WORDSIZE_64, __BYTE_ORDER__]):
 	Endian and wordsize adjustments for bit fields.
 	* src/sassim.cpp (logNodeType[__WORDSIZE_64, __BYTE_ORDER__]):
 	Endian and wordsize adjustments for bit fields.

 	*src/sasconf.h [__LP64__]: If not already defined,
 	define __WORDSIZE_64/32 from __LP64__.

Signed-off-by: Steven Munroe <munroesj52@gmail.com>